printf format checks

+++ modules/appl/xslog/src/xslog_simple.c
@@ -34,7 +34,7 @@ void xvslog(int priority, const char *format, va_list ap)
 {
   xtime_ms_t ts = xtime_ms();
 
-  xprintf("%5d.%03d: ", ts / 1000, ts % 1000);
+  xprintf("%5d.%03d: ", (int)(ts / 1000), (int)(ts % 1000));
   xvprintf(format, ap);
   xprintf("\n");
 }
